Changes from Code review
