--- lib/ui/theme.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class AppTheme {
-  AppTheme._();
-
-  //colors
-  static const primary = Color(0xFF0097A7);
-
-  static const labelColor = Colors.white;
-
-  //styles
-  static const labelTextStyle = TextStyle(fontSize: 12.0, color: labelColor);
-}
